ui refactoring
