+++ PYME/DSView/modules/measure/filterPan.py
@@ -28,7 +28,7 @@ import sys
 import PYME.ui.manualFoldPanel as afp
 
 from PYME.ui import histLimits
-from PYME.LMVis import editFilterDialog
+from PYME.ui import editFilterDialog
 
 def CreateFilterPane(panel, mapping, pipeline, visFr):
     pane = FilterPane(panel, mapping, pipeline, visFr)
