+++ PYME/ui/editFilterDialog.py
